Сделаны правки
